requirements.txt
torch==1.12.1
ray
segmentation-models-pytorch
geojson==2.5.0
matplotlib==3.5.1
networkx==2.7.1
numpy==1.23.4
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
opencv-python-headless==4.5.5.64
openpyxl==3.0.9
openslide-python
openTSNE==0.7.1
pandas==1.5.1
Pillow==8.2.0
requests==2.25.1
scikit-image==0.19.2
albumentations
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
tqdm==4.64.1
torchstain